feat(text): opt font style
