+++ packages/core/src/2d/index.ts
@@ -2,6 +2,7 @@ export { SpriteMaskInteraction } from "./enums/SpriteMaskInteraction";
 export { SpriteMaskLayer } from "./enums/SpriteMaskLayer";
 export { TextHorizontalAlignment, TextVerticalAlignment } from "./enums/TextAlignment";
 export { OverflowMode } from "./enums/TextOverflow";
+export { FontStyle } from "./enums/FontStyle";
 export { SpriteAtlas } from "./atlas/SpriteAtlas";
 export { DynamicTextAtlasManager } from "./dynamic-atlas/DynamicTextAtlasManager";
 export * from "./sprite/index";
